Greatly improve accuracy of pricing.  Partial items will now be sold with the proper price.

+++ src/main/java/regalowl/hyperconomy/hyperobject/Enchant.java
@@ -77,15 +77,15 @@ public class Enchant extends BasicObject implements HyperObject {
 	}
 
 	@Override
-	public double getBuyPrice(int amount) {
+	public double getBuyPrice(double amount) {
 		return getBuyPrice(EnchantmentClass.DIAMOND) * amount;
 	}
 	@Override
-	public double getSellPrice(int amount) {
+	public double getSellPrice(double amount) {
 		return getSellPrice(EnchantmentClass.DIAMOND) * amount;
 	}
 	@Override
-	public double getSellPrice(int amount, HyperPlayer hp) {
+	public double getSellPrice(double amount, HyperPlayer hp) {
 		return getSellPrice(EnchantmentClass.DIAMOND, hp) * amount;
 	}
 	@Override
